lib: move setupAllowedFlags() into per_thread.js
Because most of its code (the getter) has nothing to do with
the actual bootstrapping and it is run per-thread.
